refactor: remove `isFindInputValid` method
Refactor: remove redundant `isFindInputValid` method from `InputValidator.java`.

`isFindInputValid()` is used to validate whether the `find` command syntax is valid.

There is already a `isCheckInputValid()` method which does essentially the same thing, just with an extra check for the `int`.

It is better to maintain just one of such validator method as it will maintain a single level of abstraction. Changing such a method in the future would just require us to change one method, ensuring that there is fewer points of errors.

Let's:
- remove `isFindInputValid()` and rename `isCheckInputValid()` to a more general name: `isSingleInputValid()`